rename file and remove try and catch
